second try

+++ Labs/Lab5/Accounts/Account.cs
@@ -3,9 +3,9 @@ namespace Lab5.Accounts
     public abstract class Account
     {
         public int Id { get; }
-        public int Balance { get; set; }
+        public double Balance { get; set; }
 
-        public Account(int id, int balance)
+        public Account(int id, double balance)
         {
             Id = id;
             Balance = balance;
